requirements.txt
Flask
requests
cloudinary
python-dotenv
gunicorn